requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.26
reportlab>=4.0
qrcode>=7.4.2
Pillow>=10.0
